User commit

+++ src/main/java/com/example/demo/controller/ValidationController.java
@@ -16,7 +16,7 @@ public class ValidationController{
         return ser.postData(valid);
     }
     @GetMapping("/get_validate/{id}")
-    public ValidationEntity getidval(@PathVariable int id){
+    public ValidationEntity getidval(@PathVariable Long id){
         return ser.getData(id);
     }
 }
