+++ src/main/java/com/example/demo/service/ValidationService.java
@@ -2,5 +2,5 @@ package com.example.demo.service;
 import com.example.demo.entity.ValidationEntity;
 public interface ValidationService{
     ValidationEntity postData(ValidationEntity valid);
-    ValidationEntity getData(Integer id);
+    ValidationEntity getData(Long id);
 }
